Add files via upload
Project Summary:
This project is designed with an impressive and innovative approach, utilizing the capabilities of modern technology. Key components such as sensors and development boards are crucial for the success of the project.

Main Components and Features:

Raspberry Pi Pico: Provides flexibility and robust processing power, essential for the effective setup and operation of the burglar alarm circuit.
PIR Motion Sensor and HC-SR04 Ultrasonic Distance Sensor: The integration of these sensors enhances the project's detection capabilities, enabling more precise outcomes.
RGB LED, Buzzer, and Other Components: Offer a user-friendly interface and an interactive experience.
Testing and Simulation:

wokwi.com Platform: Offers the opportunity to test and simulate the project in an environment similar to real-world conditions. The project's simulation environment is available on this platform. For details, visit: https://wokwi.com/projects/366451217488434177

Conclusion:
This project aims to provide an effective and innovative solution by fully leveraging the possibilities of modern technology.